+++ sbatch_L2.sh
@@ -1,13 +1,13 @@
 #!/bin/bash -e
 #SBATCH --job-name=L2
-#SBATCH --output=/lustre/scratch/client/vinai/users/hoanglh88/IGMC/slurms/May15_L2.out
-#SBATCH --error=/lustre/scratch/client/vinai/users/hoanglh88/IGMC/slurms/May15_L2.err
+#SBATCH --output=/lustre/scratch/client/vinai/users/hoanglh88/IGMC/slurms/May17_L2.out
+#SBATCH --error=/lustre/scratch/client/vinai/users/hoanglh88/IGMC/slurms/May17_L2.err
 #SBATCH --partition=research
-#SBATCH --gres=gpu:2              # gpu count
+#SBATCH --gres=gpu:4              # gpu count
 #SBATCH --ntasks=1                 # total number of tasks across all nodes
 #SBATCH --nodes=1                  # node count
 #SBATCH --mem=256G
-#SBATCH --cpus-per-task=4          # cpu-cores per task (>1 if multi-threaded tasks)
+#SBATCH --cpus-per-task=8          # cpu-cores per task (>1 if multi-threaded tasks)
 
 
 export HTTP_PROXY=http://proxytc.vingroup.net:9090/ \
@@ -33,49 +33,54 @@ srun --container-image=/lustre/scratch/client/vinai/users/hoanglh88/utils/dc-min
 
      cd /hoanglh88/IGMC
 
-     python -m regularization.train --dataset flixster --pe_dim 40 -g 0 --n_perm_graphs 6 --metric L2 &\
-     python -m regularization.train --dataset flixster --pe_dim 86 -g 0 --n_perm_graphs 6 --metric L2
-     python -m regularization.train --dataset yahoo_music --pe_dim 40 -g 0 --n_perm_graphs 6 --metric L2 &\
-     python -m regularization.train --dataset yahoo_music --pe_dim 140 -g 0 --n_perm_graphs 6 --metric L2
+     python bot.py --scenario  1 -g 0 --metric L2 --mixer trans_encoder --dataset yahoo_music --mode coop &\
+     python bot.py --scenario  2 -g 1 --metric L2 --mixer trans_encoder --dataset yahoo_music --mode coop &\
+     python bot.py --scenario  3 -g 2 --metric L2 --mixer trans_encoder --dataset yahoo_music --mode coop &\
+     python bot.py --scenario  4 -g 3 --metric L2 --mixer trans_encoder --dataset yahoo_music --mode coop &\
+     python bot.py --scenario  5 -g 0 --metric L2 --mixer trans_encoder --dataset yahoo_music --mode coop &\
+     python bot.py --scenario  6 -g 1 --metric L2 --mixer trans_encoder --dataset yahoo_music --mode coop &\
+     python bot.py --scenario  7 -g 2 --metric L2 --mixer trans_encoder --dataset yahoo_music --mode coop &\
+     python bot.py --scenario  8 -g 3 --metric L2 --mixer trans_encoder --dataset yahoo_music --mode coop
+     python bot.py --scenario  9 -g 0 --metric L2 --mixer trans_encoder --dataset yahoo_music --mode coop &\
+     python bot.py --scenario 10 -g 1 --metric L2 --mixer trans_encoder --dataset yahoo_music --mode coop &\ 
+     python bot.py --scenario 11 -g 2 --metric L2 --mixer trans_encoder --dataset yahoo_music --mode coop &\
+     python bot.py --scenario 12 -g 3 --metric L2 --mixer trans_encoder --dataset yahoo_music --mode coop &\
+     python bot.py --scenario 13 -g 0 --metric L2 --mixer trans_encoder --dataset yahoo_music --mode coop &\
+     python bot.py --scenario 14 -g 1 --metric L2 --mixer trans_encoder --dataset yahoo_music --mode coop &\
+     python bot.py --scenario 15 -g 2 --metric L2 --mixer trans_encoder --dataset yahoo_music --mode coop &\
+     python bot.py --scenario 16 -g 3 --metric L2 --mixer trans_encoder --dataset yahoo_music --mode coop
 
-     python bot.py --scenario 1 -g 0 --metric L2 --mixer trans_encoder --dataset flix_yah &\
-     python bot.py --scenario 2 -g 0 --metric L2 --mixer trans_encoder --dataset flix_yah
-     python bot.py --scenario 3 -g 0 --metric L2 --mixer trans_encoder --dataset flix_yah &\
-     python bot.py --scenario 4 -g 0 --metric L2 --mixer trans_encoder --dataset flix_yah 
-     python bot.py --scenario 5 -g 0 --metric L2 --mixer trans_encoder --dataset flix_yah &\
-     python bot.py --scenario 6 -g 0 --metric L2 --mixer trans_encoder --dataset flix_yah 
-     python bot.py --scenario 7 -g 0 --metric L2 --mixer trans_encoder --dataset flix_yah &\
-     python bot.py --scenario 8 -g 0 --metric L2 --mixer trans_encoder --dataset flix_yah
+     python bot.py --scenario  1 -g 0 --metric L2 --mixer trans_encoder --dataset flixster --mode coop &\
+     python bot.py --scenario  2 -g 1 --metric L2 --mixer trans_encoder --dataset flixster --mode coop &\
+     python bot.py --scenario  3 -g 2 --metric L2 --mixer trans_encoder --dataset flixster --mode coop &\
+     python bot.py --scenario  4 -g 3 --metric L2 --mixer trans_encoder --dataset flixster --mode coop &\
+     python bot.py --scenario  5 -g 0 --metric L2 --mixer trans_encoder --dataset flixster --mode coop &\
+     python bot.py --scenario  6 -g 1 --metric L2 --mixer trans_encoder --dataset flixster --mode coop &\
+     python bot.py --scenario  7 -g 2 --metric L2 --mixer trans_encoder --dataset flixster --mode coop &\
+     python bot.py --scenario  8 -g 3 --metric L2 --mixer trans_encoder --dataset flixster --mode coop
+     python bot.py --scenario  9 -g 0 --metric L2 --mixer trans_encoder --dataset flixster --mode coop &\
+     python bot.py --scenario 10 -g 1 --metric L2 --mixer trans_encoder --dataset flixster --mode coop &\ 
+     python bot.py --scenario 11 -g 2 --metric L2 --mixer trans_encoder --dataset flixster --mode coop &\
+     python bot.py --scenario 12 -g 3 --metric L2 --mixer trans_encoder --dataset flixster --mode coop &\
+     python bot.py --scenario 13 -g 0 --metric L2 --mixer trans_encoder --dataset flixster --mode coop &\
+     python bot.py --scenario 14 -g 1 --metric L2 --mixer trans_encoder --dataset flixster --mode coop &\
+     python bot.py --scenario 15 -g 2 --metric L2 --mixer trans_encoder --dataset flixster --mode coop &\
+     python bot.py --scenario 16 -g 3 --metric L2 --mixer trans_encoder --dataset flixster --mode coop
 
-     python bot.py --scenario 1 -g 0 --metric L2 --mixer hyper_mixer --dataset flix_yah &\
-     python bot.py --scenario 2 -g 0 --metric L2 --mixer hyper_mixer --dataset flix_yah
-     python bot.py --scenario 3 -g 0 --metric L2 --mixer hyper_mixer --dataset flix_yah &\
-     python bot.py --scenario 4 -g 0 --metric L2 --mixer hyper_mixer --dataset flix_yah 
-     python bot.py --scenario 5 -g 0 --metric L2 --mixer hyper_mixer --dataset flix_yah &\
-     python bot.py --scenario 6 -g 0 --metric L2 --mixer hyper_mixer --dataset flix_yah 
-     python bot.py --scenario 7 -g 0 --metric L2 --mixer hyper_mixer --dataset flix_yah &\
-     python bot.py --scenario 8 -g 0 --metric L2 --mixer hyper_mixer --dataset flix_yah
-
-
-
-     python -m regularization.train --dataset douban --pe_dim 40 -g 0 --n_perm_graphs 6 --metric L2 &\
-     python -m regularization.train --dataset douban --pe_dim 115 -g 1 --n_perm_graphs 6 --metric L2 
-
-     python bot.py --scenario 9 -g 0 --metric L2 --mixer trans_encoder --dataset douban &\
-     python bot.py --scenario 10 -g 1 --metric L2 --mixer trans_encoder --dataset douban
-     python bot.py --scenario 11 -g 0 --metric L2 --mixer trans_encoder --dataset douban &\
-     python bot.py --scenario 12 -g 1 --metric L2 --mixer trans_encoder --dataset douban 
-     python bot.py --scenario 13 -g 0 --metric L2 --mixer trans_encoder --dataset douban &\
-     python bot.py --scenario 14 -g 1 --metric L2 --mixer trans_encoder --dataset douban 
-     python bot.py --scenario 15 -g 0 --metric L2 --mixer trans_encoder --dataset douban &\
-     python bot.py --scenario 16 -g 1 --metric L2 --mixer trans_encoder --dataset douban
-
-     python bot.py --scenario 9 -g 0 --metric L2 --mixer hyper_mixer --dataset douban &\
-     python bot.py --scenario 10 -g 1 --metric L2 --mixer hyper_mixer --dataset douban
-     python bot.py --scenario 11 -g 0 --metric L2 --mixer hyper_mixer --dataset douban &\
-     python bot.py --scenario 12 -g 1 --metric L2 --mixer hyper_mixer --dataset douban 
-     python bot.py --scenario 13 -g 0 --metric L2 --mixer hyper_mixer --dataset douban &\
-     python bot.py --scenario 14 -g 1 --metric L2 --mixer hyper_mixer --dataset douban 
-     python bot.py --scenario 15 -g 0 --metric L2 --mixer hyper_mixer --dataset douban &\
-     python bot.py --scenario 16 -g 1 --metric L2 --mixer hyper_mixer --dataset douban
+     python bot.py --scenario  1 -g 0 --metric L2 --mixer trans_encoder --dataset douban --mode coop &\
+     python bot.py --scenario  2 -g 1 --metric L2 --mixer trans_encoder --dataset douban --mode coop &\
+     python bot.py --scenario  3 -g 2 --metric L2 --mixer trans_encoder --dataset douban --mode coop &\
+     python bot.py --scenario  4 -g 3 --metric L2 --mixer trans_encoder --dataset douban --mode coop &\
+     python bot.py --scenario  5 -g 0 --metric L2 --mixer trans_encoder --dataset douban --mode coop &\
+     python bot.py --scenario  6 -g 1 --metric L2 --mixer trans_encoder --dataset douban --mode coop &\
+     python bot.py --scenario  7 -g 2 --metric L2 --mixer trans_encoder --dataset douban --mode coop &\
+     python bot.py --scenario  8 -g 3 --metric L2 --mixer trans_encoder --dataset douban --mode coop
+     python bot.py --scenario  9 -g 0 --metric L2 --mixer trans_encoder --dataset douban --mode coop &\
+     python bot.py --scenario 10 -g 1 --metric L2 --mixer trans_encoder --dataset douban --mode coop &\ 
+     python bot.py --scenario 11 -g 2 --metric L2 --mixer trans_encoder --dataset douban --mode coop &\
+     python bot.py --scenario 12 -g 3 --metric L2 --mixer trans_encoder --dataset douban --mode coop &\
+     python bot.py --scenario 13 -g 0 --metric L2 --mixer trans_encoder --dataset douban --mode coop &\
+     python bot.py --scenario 14 -g 1 --metric L2 --mixer trans_encoder --dataset douban --mode coop &\
+     python bot.py --scenario 15 -g 2 --metric L2 --mixer trans_encoder --dataset douban --mode coop &\
+     python bot.py --scenario 16 -g 3 --metric L2 --mixer trans_encoder --dataset douban --mode coop
      "